requirements.txt
# Web Framework & API
fastapi
uvicorn
pydantic>=2.0
pydantic-settings

# LangChain Core & Agents
langchain
langchain-community

# AI Model Providers & Embeddings
langchain-openai
langchain-google-genai
langchain-huggingface
sentence-transformers # Required for HuggingFace embeddings

# Database & Storage
pymongo
langchain-mongodb
langgraph-checkpoint-sqlite
sqlalchemy # Required for SQLChatMessageHistory

# Security & Authentication
passlib[bcrypt]
python-jose[cryptography]
python-multipart

# Utilities
python-dotenv
loguru
